Porting of Last game assets
In this update I imported a lot menu and button codes and assets from the last semsters game to see if I could use it for anything. For now they dont do anything though. I also worked on the starting screen making the button a little cleaner and have the the basis for a bject that will write the game version at the start screen for testing